Sign in the user after account creation
